refactor(macos): extract Command Palette from AppModel into AppModel+CommandPalette.swift
Continues the AppModel god-object split. Relocates the ⌘K Command Palette — the
~190-line action catalog (paletteActions), palette execution + feature-tour
entry, and the recents/pinned management (record/pin/unpin/toggle + JSON
persistence) plus the PaletteAction model — into a new `extension AppModel`
file.

Pure relocation: the 313 moved lines are byte-identical to what was removed
(verified per range). AppModel.swift drops 4,964 -> 4,651 lines. The palette's
six instance stored props (the modal flags + paletteRecentActionIds /
palettePinnedActionIds) stay on the main class under their existing MARKs;
stored properties can't live in an extension.

Four access promotions, all needed by the moved methods:
- paletteRecentActionIdsKey, palettePinnedActionIdsKey (private static -> internal):
  the UserDefaults keys the moved persist / record methods reference.
- paletteRecentActionIds, palettePinnedActionIds (private(set) -> internal setter):
  the recents/pinned lists the moved record / pin / unpin methods write.
persist + the static decode/encode/append helpers move with the domain and stay
self-contained.

Verified: clean rebuild (rm -rf macos/.build && swift build) + swift test
(889 tests) green.
